+++ controllers/operator/mongodbstandalone_controller_test.go
@@ -100,7 +100,7 @@ func TestStandaloneAuthenticationOwnedByOpsManager(t *testing.T) {
 	reconciler, client := defaultStandaloneReconciler(st)
 	reconciler.omConnectionFactory = func(context *om.OMContext) om.Connection {
 		context.Version = versionutil.OpsManagerVersion{
-			VersionString: "4.4.2",
+			VersionString: "5.0.0",
 		}
 		conn := om.NewEmptyMockedOmConnection(context)
 		return conn
@@ -124,7 +124,7 @@ func TestStandaloneAuthenticationOwnedByOperator(t *testing.T) {
 	reconciler, client := defaultStandaloneReconciler(st)
 	reconciler.omConnectionFactory = func(context *om.OMContext) om.Connection {
 		context.Version = versionutil.OpsManagerVersion{
-			VersionString: "4.4.2",
+			VersionString: "5.0.0",
 		}
 		conn := om.NewEmptyMockedOmConnection(context)
 		return conn